Fix typo bug on Int64Serializer helper class.

+++ src/ZoneTree/Serializers/Int64Serializer.cs
@@ -4,7 +4,7 @@ public sealed class Int64Serializer : ISerializer<long>
 {
     public long Deserialize(byte[] bytes)
     {
-        return BitConverter.ToInt32(bytes);
+        return BitConverter.ToInt64(bytes);
     }
 
     public byte[] Serialize(in long entry)
